test(flashcard-session): repoint tests to renamed module paths
Collateral fallout from the study-session -> flashcard-session rename:
27 test files imported directly from the old @/components/study-session
path, plus session-core.test.js exercised the now-renamed
useStudySessionCore (-> useSessionQueue, session-core.ts ->
session-queue.ts). Also drops the two study_mode assertions left over
from the deleted DeckConfig field. No new coverage — pure import/name
repoint, all data-testid and i18n key expectations unchanged.
