+++ tests/integration/components/modals/study-session/rating-buttons.test.js
@@ -2,8 +2,8 @@ import { describe, test, expect, beforeEach } from 'vite-plus/test'
 import { mount } from '@vue/test-utils'
 import { ref } from 'vue'
 import { Rating } from 'ts-fsrs'
-import RatingButtons from '@/components/study-session/session-flashcard/rating-buttons/index.vue'
-import { PrimedGradeKey } from '@/components/study-session/session-flashcard/primed-grade-context'
+import RatingButtons from '@/components/flashcard-session/flashcard/rating-buttons/index.vue'
+import { PrimedGradeKey } from '@/components/flashcard-session/flashcard/primed-grade-context'
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
