+++ tests/integration/views/deck/deck-hero/study-button.test.js
@@ -4,7 +4,7 @@ import { defineComponent, h, useAttrs } from 'vue'
 
 const { mockStartStudy } = vi.hoisted(() => ({ mockStartStudy: vi.fn() }))
 
-vi.mock('@/components/study-session/composables/study-modal', () => ({
+vi.mock('@/components/flashcard-session/composables/study-modal', () => ({
   useStudyModal: () => ({ start: mockStartStudy })
 }))
