+++ tests/unit/composables/study-session/card-edit.test.js
@@ -1,6 +1,6 @@
 import { describe, test, expect, vi, beforeEach } from 'vite-plus/test'
 import { ref, nextTick } from 'vue'
-import { useCardEdit } from '@/components/study-session/composables/card-edit'
+import { useCardEdit } from '@/components/flashcard-session/composables/card-edit'
 
 // ── Hoisted mocks ─────────────────────────────────────────────────────────────
