+++ tests/unit/composables/study-session/session-queue.test.js
@@ -1,8 +1,8 @@
 import { describe, test, expect, vi, beforeEach, afterEach } from 'vite-plus/test'
 import { nextTick } from 'vue'
 import { Rating } from 'ts-fsrs'
-import { useStudySessionCore } from '@/components/study-session/composables/session-core'
-import { readPersistedSession } from '@/components/study-session/composables/session-persistence'
+import { useSessionQueue } from '@/components/flashcard-session/composables/session-queue'
+import { readPersistedSession } from '@/components/flashcard-session/composables/session-persistence'
 import { card } from '../../../fixtures/card'
 
 const { saveReviewMock } = vi.hoisted(() => ({
@@ -104,7 +104,7 @@ describe('session-core — shuffle: true', () => {
 
     let found_review_before_new = false
     for (let i = 0; i < RUNS; i++) {
-      const session = useStudySessionCore({ study_all_cards: true, shuffle: true })
+      const session = useSessionQueue({ study_all_cards: true, shuffle: true })
       session.setCards(all_cards)
 
       const result_ids = session.cards.value.map((c) => c.id)
@@ -127,7 +127,7 @@ describe('session-core — shuffle: true', () => {
     const new_cards = Array.from({ length: 3 }, () => makeNewCard())
     const review_cards = Array.from({ length: 4 }, () => makeReviewCard())
 
-    const session = useStudySessionCore({ study_all_cards: true, shuffle: true })
+    const session = useSessionQueue({ study_all_cards: true, shuffle: true })
     session.setCards([...new_cards, ...review_cards])
 
     expect(session.cards.value).toHaveLength(7)
@@ -137,7 +137,7 @@ describe('session-core — shuffle: true', () => {
     const cards = [makeNewCard(), makeNewCard(), makeReviewCard(), makeReviewCard()]
     const expected_ids = new Set(cards.map((c) => c.id))
 
-    const session = useStudySessionCore({ study_all_cards: true, shuffle: true })
+    const session = useSessionQueue({ study_all_cards: true, shuffle: true })
     session.setCards(cards)
 
     const result_ids = new Set(session.cards.value.map((c) => c.id))
@@ -152,7 +152,7 @@ describe('session-core — shuffle: false', () => {
     const cards = [makeNewCard(), makeReviewCard(), makeNewCard(), makeReviewCard()]
     const original_ids = cards.map((c) => c.id)
 
-    const session = useStudySessionCore({ study_all_cards: true, shuffle: false })
+    const session = useSessionQueue({ study_all_cards: true, shuffle: false })
     session.setCards(cards)
 
     const result_ids = session.cards.value.map((c) => c.id)
@@ -164,7 +164,7 @@ describe('session-core — shuffle: false', () => {
     const cards = [makeNewCard(), makeReviewCard(), makeNewCard(), makeReviewCard()]
     const original_ids = cards.map((c) => c.id)
 
-    const session = useStudySessionCore({ study_all_cards: true, shuffle: false })
+    const session = useSessionQueue({ study_all_cards: true, shuffle: false })
     session.setCards(cards)
 
     expect(session.cards.value.map((c) => c.id)).toEqual(original_ids)
@@ -175,19 +175,19 @@ describe('session-core — shuffle: false', () => {
 
 describe('session-core — setCards', () => {
   test('study_all_cards:true includes all cards', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard(), makeNotDueCard()])
     expect(session.cards.value).toHaveLength(2)
   })
 
   test('study_all_cards:false filters out cards with a future due date', () => {
-    const session = useStudySessionCore({ study_all_cards: false })
+    const session = useSessionQueue({ study_all_cards: false })
     session.setCards([makeNewCard(), makeNotDueCard()])
     expect(session.cards.value).toHaveLength(1)
   })
 
   test('empty deck sets mode to completed immediately', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([])
     expect(session.mode.value).toBe('completed')
   })
@@ -200,7 +200,7 @@ describe('session-core — updateConfig', () => {
     const cards = [makeNewCard(), makeReviewCard()]
     const original_ids = cards.map((c) => c.id)
 
-    const session = useStudySessionCore({ study_all_cards: true, shuffle: false })
+    const session = useSessionQueue({ study_all_cards: true, shuffle: false })
     session.setCards(cards)
 
     // Confirm un-shuffled order first
@@ -212,7 +212,7 @@ describe('session-core — updateConfig', () => {
   })
 
   test('updateConfig does not call _processCards when no cards have been set', () => {
-    const session = useStudySessionCore({ study_all_cards: false })
+    const session = useSessionQueue({ study_all_cards: false })
     session.updateConfig({ study_all_cards: true })
     expect(session.cards.value).toHaveLength(0)
   })
@@ -223,20 +223,20 @@ describe('session-core — updateConfig', () => {
 describe('session-core — member preference seeding [obligation]', () => {
   test('show_all_ratings seeds from member_store.preferences.study.show_all_ratings [obligation]', () => {
     mockMemberStore.preferences.study.show_all_ratings = false
-    const session = useStudySessionCore()
+    const session = useSessionQueue()
     expect(session.show_all_ratings.value).toBe(false)
   })
 
   test('show_all_ratings is a local ref, toggling it does not write back to the store [obligation]', () => {
     mockMemberStore.preferences.study.show_all_ratings = true
-    const session = useStudySessionCore()
+    const session = useSessionQueue()
     session.show_all_ratings.value = false
     expect(mockMemberStore.preferences.study.show_all_ratings).toBe(true)
   })
 
   test('FSRS is seeded with request_retention = desired_retention / 100 [obligation]', () => {
     mockMemberStore.preferences.study.desired_retention = 82
-    useStudySessionCore()
+    useSessionQueue()
     expect(generatorParametersMock).toHaveBeenCalledWith(
       expect.objectContaining({ request_retention: 0.82 })
     )
@@ -247,7 +247,7 @@ describe('session-core — member preference seeding [obligation]', () => {
     // silently disabling learning/relearning steps regardless of member prefs.
     mockMemberStore.preferences.study.learning_steps = ['1m', '10m', '1d']
     mockMemberStore.preferences.study.relearning_steps = ['1m', '10m']
-    useStudySessionCore()
+    useSessionQueue()
     expect(generatorParametersMock).toHaveBeenCalledWith(
       expect.objectContaining({
         learning_steps: ['1m', '10m', '1d'],
@@ -261,7 +261,7 @@ describe('session-core — member preference seeding [obligation]', () => {
 
 describe('session-core — reviewCard (no grade)', () => {
   test('no-grade review marks the active card as passed and advances', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -274,7 +274,7 @@ describe('session-core — reviewCard (no grade)', () => {
   })
 
   test('no-grade review on the last card sets mode to completed', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
 
     session.reviewCard()
@@ -283,7 +283,7 @@ describe('session-core — reviewCard (no grade)', () => {
   })
 
   test('no-grade review does NOT call save_review_mutation', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
 
     session.reviewCard()
@@ -292,7 +292,7 @@ describe('session-core — reviewCard (no grade)', () => {
   })
 
   test('reviewCard with no active_card is a no-op', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     // No setCards — active_card is undefined
     session.reviewCard()
 
@@ -305,7 +305,7 @@ describe('session-core — reviewCard (no grade)', () => {
 
 describe('session-core — reviewCard (with grade)', () => {
   test('grade=Good marks the active card as passed', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -315,7 +315,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('grade=Again marks the active card as failed', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -325,7 +325,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('grade=Hard marks the active card as passed (not failed)', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -335,7 +335,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('graded review advances to the next unreviewed card', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -345,7 +345,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('graded review on the last card sets mode to completed', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
 
     session.reviewCard(Rating.Good)
@@ -354,7 +354,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('graded review calls save_review_mutation with card_id and deck_id', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 42, deck_id: 7 })
     session.setCards([c])
 
@@ -366,7 +366,7 @@ describe('session-core — reviewCard (with grade)', () => {
   })
 
   test('graded review does NOT call save when card has no id', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     // id falsy — mutation should be skipped
     const c = makeNewCard({ id: 0 })
     session.setCards([c])
@@ -382,7 +382,7 @@ describe('session-core — reviewCard (with grade)', () => {
 describe('session-core — reviewCard save failure [obligation]', () => {
   test('resolves (does not reject) when save_review_mutation.mutateAsync rejects, since the caller awaits it as a floating promise [obligation]', async () => {
     saveReviewMock.mockRejectedValueOnce(new Error('network error'))
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -391,7 +391,7 @@ describe('session-core — reviewCard save failure [obligation]', () => {
 
   test('shows a panel notice with a Refresh action when the save fails [obligation]', async () => {
     saveReviewMock.mockRejectedValueOnce(new Error('network error'))
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -416,7 +416,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
     // The central invariant: the before-interval must be the PRIOR scheduled_days,
     // captured at the moment of review, before FSRS overwrites card.review.
     const prior_scheduled_days = 10
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = card.one({
       traits: 'with_due_review',
       overrides: {
@@ -445,7 +445,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('is_new is true for a card with no prior review (reps=0) [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -455,7 +455,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('is_new is false when card has prior reviews (reps > 0) [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeReviewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -465,7 +465,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('reviewCard() with no grade (auto-pass) does NOT push a result [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1 })
     session.setCards([c])
 
@@ -475,7 +475,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('reviewCard(grade) pushes exactly one result per graded review [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard({ id: 1, deck_id: 1 }), makeNewCard({ id: 2, deck_id: 1 })]
     session.setCards([c1, c2])
 
@@ -486,7 +486,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('setCards clears results from a previous session [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard({ id: 1, deck_id: 1 }), makeNewCard({ id: 2, deck_id: 1 })]
     session.setCards([c1])
     session.reviewCard(Rating.Good)
@@ -498,7 +498,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('passed = true for grade Hard [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -508,7 +508,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('passed = true for grade Easy [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c = makeNewCard({ id: 1, deck_id: 1 })
     session.setCards([c])
 
@@ -518,7 +518,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
   })
 
   test('passed = false for grade Again [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard({ id: 1, deck_id: 1 }), makeNewCard({ id: 2, deck_id: 1 })]
     session.setCards([c1, c2])
 
@@ -532,7 +532,7 @@ describe('session-core — results (CardReviewResult capture)', () => {
 
 describe('session-core — computed stats', () => {
   test('num_correct counts only passed cards', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
 
@@ -543,7 +543,7 @@ describe('session-core — computed stats', () => {
   })
 
   test('reviewed_count counts both passed and failed cards', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
 
@@ -554,7 +554,7 @@ describe('session-core — computed stats', () => {
   })
 
   test('current_index returns the index of the active card', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -566,7 +566,7 @@ describe('session-core — computed stats', () => {
   })
 
   test('current_index returns cards.length when no active card', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
 
     session.reviewCard()
@@ -575,14 +575,14 @@ describe('session-core — computed stats', () => {
   })
 
   test('remaining_due_count is 0 when study_all_cards is true', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard(), makeReviewCard()])
 
     expect(session.remaining_due_count.value).toBe(0)
   })
 
   test('remaining_due_count reflects total due minus already-reviewed count', () => {
-    const session = useStudySessionCore({ study_all_cards: false })
+    const session = useSessionQueue({ study_all_cards: false })
     // Two due review cards — remaining starts at 2
     session.setCards([makeReviewCard(), makeReviewCard()])
 
@@ -590,7 +590,7 @@ describe('session-core — computed stats', () => {
   })
 
   test('remaining_due_count is 0 when no due cards in deck', () => {
-    const session = useStudySessionCore({ study_all_cards: false })
+    const session = useSessionQueue({ study_all_cards: false })
     // Only not-due cards — they don't show up at all
     session.setCards([makeNotDueCard()])
 
@@ -602,7 +602,7 @@ describe('session-core — computed stats', () => {
 
 describe('session-core — dropCard [obligation]', () => {
   test('removing the active card advances active_card to the next unreviewed card [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
 
@@ -614,7 +614,7 @@ describe('session-core — dropCard [obligation]', () => {
   })
 
   test('dropping the active card removes it from cards [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -625,7 +625,7 @@ describe('session-core — dropCard [obligation]', () => {
   })
 
   test('dropping the active card also removes it from the due pool (remaining_due_count decreases) [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: false })
+    const session = useSessionQueue({ study_all_cards: false })
     const [c1, c2] = [makeReviewCard(), makeReviewCard()]
     session.setCards([c1, c2])
 
@@ -637,7 +637,7 @@ describe('session-core — dropCard [obligation]', () => {
   })
 
   test('removing the last remaining card sets mode to completed [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const c1 = makeNewCard()
     session.setCards([c1])
 
@@ -647,7 +647,7 @@ describe('session-core — dropCard [obligation]', () => {
   })
 
   test('removing a non-active card leaves active_card unchanged [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
 
@@ -660,7 +660,7 @@ describe('session-core — dropCard [obligation]', () => {
   })
 
   test('removing a non-active card removes it from cards but leaves the rest', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
 
@@ -674,7 +674,7 @@ describe('session-core — dropCard [obligation]', () => {
 
 describe('session-core — updateCard [obligation]', () => {
   test('patches the matching card in cards [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -684,7 +684,7 @@ describe('session-core — updateCard [obligation]', () => {
   })
 
   test('reassigns the cards array reference (shallowRef immutability) [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
     const before = session.cards.value
@@ -695,7 +695,7 @@ describe('session-core — updateCard [obligation]', () => {
   })
 
   test('leaves non-matching cards untouched (value and reference) [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
     const c2_before = session.cards.value.find((c) => c.id === c2.id)
@@ -707,7 +707,7 @@ describe('session-core — updateCard [obligation]', () => {
   })
 
   test('patches active_card when the patched card is the active one [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -719,7 +719,7 @@ describe('session-core — updateCard [obligation]', () => {
   })
 
   test('does not touch active_card when the patched card is not the active one [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
     const active_before = session.active_card.value
@@ -734,7 +734,7 @@ describe('session-core — updateCard [obligation]', () => {
 
 describe('session-core — persistence [obligation]', () => {
   test('setCards persists card_ids into sessionStorage [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
     await nextTick()
@@ -745,7 +745,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('setSessionMeta records deck_ids and config_override into the next persisted snapshot [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setSessionMeta([1, 2], { shuffle: true })
     session.setCards([makeNewCard()])
     await nextTick()
@@ -756,7 +756,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('reviewCard(grade) updates the persisted results and card_ids, not just the initial seed [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard({ id: 1, deck_id: 1 }), makeNewCard({ id: 2, deck_id: 1 })]
     session.setCards([c1, c2])
     await nextTick()
@@ -771,7 +771,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('reviewCard() with no grade also persists the resulting completed mode [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
     await nextTick()
 
@@ -782,7 +782,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('persisted results only carry CardReviewResult fields, no full card data [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard({ id: 1, deck_id: 1 })])
     await nextTick()
 
@@ -804,7 +804,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('dropCard on a non-active card persists the updated card_ids [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2, c3] = [makeNewCard(), makeNewCard(), makeNewCard()]
     session.setCards([c1, c2, c3])
     await nextTick()
@@ -816,7 +816,7 @@ describe('session-core — persistence [obligation]', () => {
   })
 
   test('dropCard on the active card persists the updated card_ids [obligation]', async () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
     await nextTick()
@@ -841,12 +841,12 @@ describe('session-core — active_card_preview [obligation]', () => {
   })
 
   test('is undefined when there is no active card [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     expect(session.active_card_preview.value).toBeUndefined()
   })
 
   test('is defined once a card is active [obligation]', () => {
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
 
     expect(session.active_card_preview.value).toBeDefined()
@@ -857,7 +857,7 @@ describe('session-core — active_card_preview [obligation]', () => {
     vi.useFakeTimers()
     vi.setSystemTime(new Date('2026-01-01T00:00:00.000Z'))
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     const [c1, c2] = [makeNewCard(), makeNewCard()]
     session.setCards([c1, c2])
 
@@ -882,7 +882,7 @@ describe('session-core — active_card_preview [obligation]', () => {
     vi.useFakeTimers()
     vi.setSystemTime(new Date('2026-01-01T00:00:00.000Z'))
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.setCards([makeNewCard()])
     const first_preview = session.active_card_preview.value
 
@@ -902,7 +902,7 @@ describe('session-core — active_card_preview [obligation]', () => {
     const c1 = makeNewCard()
     const persisted = { deck_ids: [1], card_ids: [c1.id], results: [], mode: 'studying' }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
 
     vi.setSystemTime(new Date('2026-02-01T00:00:00.000Z'))
     session.restoreCards([c1], persisted)
@@ -925,7 +925,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([c1, c2, c3], persisted)
 
     expect(session.cards.value.map((c) => c.id)).toEqual([c3.id, c1.id, c2.id])
@@ -940,7 +940,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([c1, c2, extra], persisted)
 
     expect(session.cards.value.map((c) => c.id)).toEqual([c1.id, c2.id])
@@ -956,7 +956,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     expect(() => session.restoreCards([c1], persisted)).not.toThrow()
 
     expect(session.cards.value.map((c) => c.id)).toEqual([c1.id])
@@ -981,7 +981,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     // raw only contains the unreviewed remainder — c1 must NOT need to be fetched
     session.restoreCards([c2], persisted)
 
@@ -1008,7 +1008,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([], persisted)
 
     expect(session.cards.value.find((c) => c.id === c1.id)?.state).toBe('failed')
@@ -1033,7 +1033,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([c2, c3], persisted)
 
     expect(session.active_card.value?.id).toBe(c2.id)
@@ -1047,7 +1047,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     // The only persisted id has no matching fetched card — dropped, leaving no unreviewed card.
     session.restoreCards([], persisted)
 
@@ -1064,7 +1064,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'completed'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([c1], persisted)
 
     expect(session.mode.value).toBe('completed')
@@ -1080,7 +1080,7 @@ describe('session-core — restoreCards [obligation]', () => {
       mode: 'studying'
     }
 
-    const session = useStudySessionCore({ study_all_cards: true })
+    const session = useSessionQueue({ study_all_cards: true })
     session.restoreCards([c1], persisted)
     await nextTick()
