+++ tests/unit/composables/use-study-modal.test.js
@@ -1,7 +1,7 @@
 import { describe, test, expect, vi, beforeEach } from 'vite-plus/test'
 import { flushPromises } from '@vue/test-utils'
-import { useStudyModal } from '@/components/study-session/composables/study-modal'
-import StudySession from '@/components/study-session/index.vue'
+import { useStudyModal } from '@/components/flashcard-session/composables/study-modal'
+import StudySession from '@/components/flashcard-session/index.vue'
 
 // ── Hoisted mocks ─────────────────────────────────────────────────────────────
